% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creacion_insumos.R
\name{calcular_medianas_internal}
\alias{calcular_medianas_internal}
\title{Calcula medianas con método replicado}
\usage{
calcular_medianas_internal(
  var,
  dominios,
  disenio,
  sub = F,
  env = parent.frame()
)
}
\arguments{
\item{var}{variable a estimar}

\item{dominios}{dominios para desagregar}

\item{disenio}{disenio complejo creado mediante el paquete \code{survey}}

\item{sub}{se usa para filtrar cuando el usuario lo requiere}

\item{env}{ambiente en el cual se crean algunas variable relevantes}
}
\value{
\code{vector} que contiene la variable con los conglomerados.
}
\description{
Se usa para acortar un poco el código de la función mediana
}
