% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creacion_insumos.R
\name{calcular_tabla_ratio}
\alias{calcular_tabla_ratio}
\title{Calcula ratio a partir de cierta agregación}
\usage{
calcular_tabla_ratio(var, denominador, dominios = NULL, disenio)
}
\arguments{
\item{var}{variable objetivo o numerador del ratio a calcular, dentro de un \code{dataframe}. Debe anteponerse ~}

\item{denominador}{variable denominador del ratio a calcular, dentro de un \code{dataframe}. Debe anteponerse ~}

\item{dominios}{dominios de estimación separados por signo +. Debe anteponerse ~}

\item{disenio}{disenio complejo creado mediante el paquete \code{survey}}
}
\value{
\code{dataframe} que contiene variables de agregación, variable objetivo y error estándar
}
\description{
Genera una tabla con estimaciones para una agregación determinada
}
\examples{
dc <- svydesign(ids = ~varunit, strata = ~varstrat, data = epf, weights = ~fe)
calcular_tabla_ratio(var = ~gasto_div1, denominador = ~gasto, zona+sexo, dc)
}
