% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{casen}
\alias{casen}
\title{Encuesta de Caracterización Socioeconómica Nacional 2020 - CASEN en Pandemia 2020}
\format{
dataframe compuesto por 185.437 observaciones y 6 variables
\describe{
 \item{folio}{identificador del hogar}
 \item{sexo}{sexo: 1 = hombre; 2 = mujer}
 \item{edad}{edad}
 \item{activ}{Condición de actividad económica}
 \item{ing_aut_hog}{Ingreso autónomo del hogar}
 \item{pobreza}{pobreza por ingresos: 1 = pobres extremos, 2 = pobres no extremos, 3 = no pobres}
 \item{expr}{factor de expansión regional}
 \item{estrato}{estrato de muestro}
 \item{cod_upm}{unidad primaria de muestreo}
 }
}
\source{
\url{http://observatorio.ministeriodesarrollosocial.gob.cl/encuesta-casen-en-pandemia-2020}
}
\usage{
casen
}
\description{
Datos de CASEN para el año 2020 Contiene solo algunas variables.
}
\examples{
data(casen)

}
\keyword{datasets}
