% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_internal.R
\name{check_subpop_var}
\alias{check_subpop_var}
\title{Evalúa algunos requisitos básicos de la variable de subpop}
\usage{
check_subpop_var(subpop, disenio)
}
\arguments{
\item{subpop}{string of the subpopulation filter}

\item{disenio}{complex design}
}
\value{
warning or stop
}
\description{
Evalúa si la variable es dummy
}
