% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creacion_insumos.R
\name{chequear_var_disenio}
\alias{chequear_var_disenio}
\title{Chequea que las variables de dise침o tengan el nombre correcto}
\usage{
chequear_var_disenio(data)
}
\arguments{
\item{data}{\code{dataframe} que contiene la tabla con la cual se est치 trabajando}
}
\value{
un mensaje de error
}
\description{
Comprueba que las variables de dise침o se llamen varstrat y varunit. En caso de que no se cumpla, la ejecuci칩n se detiene y se genera un error
}
