% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_internal.R
\name{concat_domains}
\alias{concat_domains}
\title{Concatena los domains y la subpoblaci贸n con signo +}
\usage{
concat_domains(domains, subpop)
}
\arguments{
\item{domains}{domains en formato string}

\item{subpop}{subpoblaci贸n ingresada por el usuario en formato string}
}
\value{
string concatenado de domains y subpoblaci贸n
}
\description{
Recibe strings con domains y subpoblaci贸n y devuelve un string concatenado con caracter +
}
