% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_internal.R
\name{convert_to_formula}
\alias{convert_to_formula}
\title{Convierte un string en una fórmula}
\usage{
convert_to_formula(var)
}
\arguments{
\item{var}{sting con el nombre de la variable}
}
\value{
variable en formato fórmula
}
\description{
Recibe un string y lo convierte en un formato de fórmula
}
