% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creacion_insumos.R
\name{crear_insumos}
\alias{crear_insumos}
\title{Crea los insumos necesarios para hacer la evaluación de estimadores de la media}
\usage{
crear_insumos(var, dominios, dise昼㸱o)
}
\arguments{
\item{var}{variable objetivo dentro de un \code{dataframe}. Debe anteponerse ~}

\item{dominios}{dominios de estimación separados por signo +. Debe anteponerse ~}

\item{diseño}{diseño complejo creado mediante el paquete \code{survey}}
}
\value{
\code{dataframe} que contiene la frecuencia de todos los dominios a evaluar
}
\description{
Genera una tabla con los siguientes insumos: media, grados de libertad,
tamaño muestral y coeficiente de variación.
}
\examples{
crear_insumos(~gastot_hd, ~zona+sexo, dc)

}
