% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creacion_insumos.R
\name{crear_insumos_tot}
\alias{crear_insumos_tot}
\title{Crea los insumos necesarios para hacer la evaluación de estimadores de totales}
\usage{
crear_insumos_tot(var, dominios = NULL, subpop = NULL, disenio, ci = F)
}
\arguments{
\item{var}{string. Variable para la cual se quiere calcular un total. Pueden introducirse varias variables separadas por un +, para obtener más totales en la tabla}

\item{dominios}{string. Dominios de estimación separados por signo +.}

\item{subpop}{integer dummy que permite filtrar por una subpoblación}

\item{disenio}{disenio complejo creado mediante el paquete \code{survey}}

\item{ci}{\code{boolean} que indica si los intervalos de confianza deben calcularse}
}
\value{
\code{dataframe} que contiene la frecuencia de todos los dominios a evaluar
}
\description{
Genera una tabla con los siguientes insumos: total expandido, grados de libertad,
tamaño muestral y coeficiente de variación. La función contempla la posibilidad de desagregar la estimación en uno o más dominios.
}
\examples{
dc <- svydesign(ids = ~varunit, strata = ~varstrat, data = epf_personas, weights = ~fe)
crear_insumos_tot(ocupado, zona+sexo, dc)
}
