% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creacion_insumos.R
\name{crear_insumos_tot_con}
\alias{crear_insumos_tot_con}
\title{Crea los insumos para el caso específico de estimaciones de suma}
\usage{
crear_insumos_tot_con(
  var,
  dominios = NULL,
  subpop = NULL,
  disenio,
  ci = F,
  ajuste_ene = F
)
}
\arguments{
\item{var}{variable objetivo dentro de un \code{dataframe}.}

\item{dominios}{dominios de estimación separados por signo +.}

\item{subpop}{integer dummy que permite filtrar por una subpoblación}

\item{disenio}{disenio complejo creado mediante el paquete \code{survey}}

\item{ci}{\code{boolean} que indica si los intervalos de confianza deben calcularse}

\item{ajuste_ene}{\code{boolean} que indica si debe aplicarse el ajuste por cambio de marco en la ENE}
}
\value{
\code{dataframe} que contiene todos los insumos necesarios para evaluar una suma
}
\description{
Genera una tabla con los siguientes insumos: suma, grados de libertad,
tamaño muestral y coeficiente de variación. La función tiene la posibilidad de utilizar una subpoblación. La variable de subpoblación debe ser dummy.
}
\examples{
dc <- svydesign(ids = ~varunit, strata = ~varstrat, data = epf_personas, weights = ~fe)
crear_insumos_suma(gastot_hd, zona+sexo, subpop = ocupado, disenio = dc)
}
