% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{create_html}
\alias{create_html}
\title{Create html table with the results of the evaluation}
\usage{
create_html(table)
}
\arguments{
\item{table}{A \code{dataframe} generated by the \code{assess} function.
It must have a class attribute indicating the evaluation scheme
(e.g., "ine.eval", "cepal2020.eval").}
}
\value{
A \code{knitr_kable} object representing the HTML table.
}
\description{
This function generates an HTML table to visualize the results of the assessment
performed by \code{assess}. It applies specific formatting and color coding
based on the evaluation schema used (INE, CEPAL 2020, CEPAL 2023, or Economics).
}
