% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creacion_insumos.R
\name{create_median}
\alias{create_median}
\title{Create the inputs to make quality evaluation for median estimations}
\usage{
create_median(
  var,
  dominios = NULL,
  subpop = NULL,
  disenio,
  ci = F,
  replicas = 10,
  ajuste_ene = F,
  anidar = F
)
}
\arguments{
\item{var}{numeric variable within the  \code{dataframe}}

\item{dominios}{domains to be estimated separated by the + character.}

\item{subpop}{integer dummy variable to filter the dataframe}

\item{disenio}{complex design created by \code{survey} package}

\item{ci}{\code{boolean} indicating if the confidence intervals must be calculated}

\item{replicas}{\code{integer} indicating the number of replicates to be used}

\item{ajuste_ene}{\code{boolean} indicating if an adjustment for the sampling-frame transition period must be used
#' @param anidar \code{boolean} Indicating if the function is wrapped inside a function, if \code{TRUE} avoid lazy eval errors}
}
\value{
\code{dataframe} that contains the inputs and all domains to be evaluated
}
\description{
\code{create_median} uses a non parametric method to generate a \code{dataframe}
with the following elements: sum, degrees of freedom, sample size and coefficient
of variation. The function allows grouping in several domains.
}
\examples{
dc <- svydesign(ids = ~varunit, strata = ~varstrat, data = epf_personas, weights = ~fe)
dc_rep <-  as.svrepdesign(dc , type = "subbootstrap", replicates=10)
create_median(gastot_hd, zona+sexo, disenio = dc)
}
