% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_internal.R
\name{create_output}
\alias{create_output}
\title{Une información de indicadores y genera tabla final}
\usage{
create_output(table, domains, gl, n, cv)
}
\arguments{
\item{table}{objeto creado con survey}

\item{domains}{listado de variables para desagregar}
}
\value{
dataframe con toda la información para estándar INE
}
\description{
Recibe los indicadores de calidad calculados previamente y los une en una tabla
}
