% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creacion_insumos.R
\name{create_prop_internal}
\alias{create_prop_internal}
\title{internal function to calculate proportion estimations}
\usage{
create_prop_internal(
  var,
  dominios = NULL,
  subpop = NULL,
  disenio,
  ci = F,
  ajuste_ene = F,
  standard_eval = T
)
}
\arguments{
\item{var}{integer dummy variable within the  \code{dataframe}}

\item{dominios}{domains to be estimated separated by the + character.}

\item{subpop}{integer dummy variable to filter the dataframe}

\item{disenio}{complex design created by \code{survey} package}

\item{ci}{\code{boolean} indicating if the confidence intervals must be calculated}

\item{ajuste_ene}{\code{boolean} indicating if an adjustment for the sampling-frame transition period must be used}

\item{standard_eval}{\code{boolean} indicating if the function is inside another function, by default it is TRUE, avoid problems with lazy eval.}
}
\value{
\code{dataframe} that contains the inputs and all domains to be evaluated
}
\description{
internal function to calculate proportion estimations
}
\examples{
dc <- svydesign(ids = ~varunit, strata = ~varstrat, data = epf_personas, weights = ~fe)
create_prop_internal(ocupado, zona+sexo, disenio = dc)
}
