% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creacion_insumos.R
\name{create_ratio}
\alias{create_ratio}
\title{\code{create_ratio} generates a \code{dataframe} with the following elements: sum,
degrees of freedom, sample size, standard error and coefficient of variation. The function allows
grouping in several domains.}
\usage{
create_ratio(
  var,
  denominador = NULL,
  dominios = NULL,
  subpop = NULL,
  disenio,
  ci = F,
  ajuste_ene = F,
  anidar = F
)
}
\arguments{
\item{var}{numeric variable within the \code{dataframe}, is the numerator of the ratio to be calculated.}

\item{denominador}{numeric variable within the \code{dataframe}, is the denominator of the ratio to be calculated. If the \code{var} parameter is dummy, it can be NULL}

\item{dominios}{domains to be estimated separated by the + character.}

\item{subpop}{integer dummy variable to filter the dataframe}

\item{disenio}{complex design created by \code{survey} package}

\item{ci}{\code{boolean} indicating if the confidence intervals must be calculated}

\item{ajuste_ene}{\code{boolean} indicating if an adjustment for the sampling-frame transition period must be used}

\item{anidar}{\code{boolean} Indicating if the function is wrapped inside a function, if \code{TRUE} avoid lazy eval errors}
}
\value{
\code{dataframe} that contains the inputs and all domains to be evaluated
}
\description{
\code{create_ratio} generates a \code{dataframe} with the following elements: sum,
degrees of freedom, sample size, standard error and coefficient of variation. The function allows
grouping in several domains.
}
