% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_internal.R
\name{create_ratio_internal}
\alias{create_ratio_internal}
\title{internal function to calculate ratios estimations}
\usage{
create_ratio_internal(
  var,
  denominador,
  dominios = NULL,
  subpop = NULL,
  disenio,
  ci = F,
  deff = F,
  ess = F,
  ajuste_ene = F,
  unweighted = F,
  rel_error = F,
  rm.na = F
)
}
\arguments{
\item{var}{numeric variable within the \code{dataframe}, is the numerator of the ratio to be calculated.}

\item{denominador}{numeric variable within the \code{dataframe}, is the denominator of the ratio to be calculated.}

\item{dominios}{domains to be estimated separated by the + character.}

\item{subpop}{integer dummy variable to filter the dataframe}

\item{disenio}{complex design created by \code{survey} package}

\item{ci}{\code{boolean} indicating if the confidence intervals must be calculated}

\item{deff}{\code{boolean} Design effect}

\item{ess}{\code{boolean} Effective sample size}

\item{ajuste_ene}{\code{boolean} indicating if an adjustment for the sampling-frame transition period must be used}

\item{unweighted}{\code{boolean} Add non weighted count if it is required}

\item{rel_error}{\code{boolean} Relative error}

\item{log_cv}{\code{boolean} logarithmic coefficient of variation}
}
\value{
\code{dataframe} that contains the inputs and all domains to be evaluated
}
\description{
internal function to calculate ratios estimations
}
