% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revision_calidad.R
\name{cuadratica}
\alias{cuadratica}
\title{Calcula el valor de una función cuadrática}
\usage{
cuadratica(p)
}
\arguments{
\item{p}{vector numérico que contiene los valores de cada una de las estimaciones de
proporción}
}
\value{
vector numérico con el cálculo realizado
}
\description{
Calcula una función diseñada por el INE, que se utiliza para hacer una comparación
con el error estándar. Si este último sobrepasa el valor de la funcíón cuadrática,
es un indicio de que el estimador tiene una alta varianza. Siguiendo la metodología
del INE, existen cálculos diferenciados para los estimadores que están por sobre y por
debajo de 0.5.
}
\examples{
cuadratica(c(0.1, 0.7, 0.5))
}
