% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{.apply_kable_styling}
\alias{.apply_kable_styling}
\title{Helper interno para temas de formato y letras}
\usage{
.apply_kable_styling(df_styled)
}
\description{
Helper interno para temas de formato y letras
}
\keyword{internal}
