% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_internal.R
\name{eclac_standard}
\alias{eclac_standard}
\title{Turn on all the indicators needed for the eclac standard}
\usage{
eclac_standard(eclac, env = parent.frame(), proportion = FALSE)
}
\arguments{
\item{eclac}{A character string specifying the eclac standard.
It should be either "eclac_2020", "eclac_2023", or "chile".}

\item{env}{The environment from which to retrieve the existing indicator values.
Defaults to the parent frame.}

\item{proportion}{A logical value indicating whether proportion indicators
should be turned on. Defaults to FALSE.}
}
\value{
A list of logical values indicating which indicators are turned on.
}
\description{
This function activates the appropriate indicators based on the
selected eclac standard and whether proportion indicators are needed.
}
