% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ELE7}
\alias{ELE7}
\title{Encuesta Longitudinal de Empresas}
\format{
dataframe with 6.592 rows y 13 columns
\describe{
 \item{rol_ficticio}{Company ID}
 \item{cod_actividad}{Economic activity}
 \item{cod_tamano}{Company size by sales}
 \item{tramo}{Inclusion range}
 \item{fe_transversal}{Cross-sectional weights}
 \item{fe_longitudinal}{Longitudinal weights}
 \item{panel}{Panel sample}
 \item{estrato}{Strata}
 \item{pob}{Finite population correction}
 \item{VA_2022}{Value added 2022, difference between production value and intermediate consumption}
 \item{VA_2022f}{VA_2022f is an adjusted version of VA_2022, where negative values are replaced with 0, while non-negative values remain unchanged.}
 \item{EMP}{Total personnel employed and hired by the company on a monthly basis}
 \item{REMP_TOTAL}{Total gross remuneration of personnel hired by the company}

 }
}
\source{
\url{https://www.ine.gob.cl/docs/default-source/encuesta-longitudinal-de-empresas/bbdd/ele-2022/base-de-datos-ele7.csv?sfvrsn=1504c58d_4&download=true}
}
\usage{
ELE7
}
\description{
ELE data for the year 2022. Contains only a few variables.
}
\examples{
data(ELE7)

}
\keyword{datasets}
