% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ELE7_n_obj}
\alias{ELE7_n_obj}
\title{Tamano muestra objetivo Encuesta Longitudinal de Empresas}
\format{
dataframe with 59 rows y 4 columns
\describe{
 \item{cod_tamano}{Company size by sales}
 \item{cod_actividad_letra}{Economic activity}
 \item{cod_actividad}{Economic activity ID}
 \item{n_obj}{Target sample size}

 }
}
\source{
\url{https://www.ine.gob.cl/docs/default-source/encuesta-longitudinal-de-empresas/metodologias/ele-2022/informe-de-calidad-ele7.pdf?sfvrsn=6ca73eb5_4}
}
\usage{
ELE7_n_obj
}
\description{
Target cross-sectional sample size ELE data for the year 2022.
}
\examples{
data(ELE7_n_obj)

}
\keyword{datasets}
