% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ene}
\alias{ene}
\title{Encuesta Nacional de Empleo - ENE. 2020-efm}
\format{
dataframe compuesto por 87.842 observaciones y 7 variables
\describe{
  \item{sexo}{sexo: 1 = hombre; 2 = mujer }
  \item{region}
  \item{cae_especifico}{condición de actividad económica}
  \item{fe}{factor de expansión}
  \item{varunit}{unidad de primera etapa}
  \item{varstrat}{estrato}
  \item{fdt}{indica si la persona pertenece a la fuerza de trabajo: 1 = pertenece; 0 = no pertenece}
  \item{ocupado}{indica de la persona está ocupada o no: 1 = ocupado; 0 = no ocupado}
  \item{desocupado}{indica de la persona está desocupada o no: 1 = desocupada; 0 = ocupada}
}
}
\source{
\url{https://www.ine.cl/estadisticas/sociales/mercado-laboral/ocupacion-y-desocupacion}
}
\usage{
ene
}
\description{
Versión reducida de la base de datos de la ENE. Contiene
algunas variables sociodemográficas y la información necesaria
para trabajar con el diseño complejo.
}
\examples{
data(ene)

}
\keyword{datasets}
