% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{enusc_2023}
\alias{enusc_2023}
\title{Encuesta Nacional Urbana de Seguridad Ciudadana 2023 - ENUSC 2023}
\format{
dataframe with 49.813 rows y 15 columns
\describe{
 \item{enc_region }{16 regions}
 \item{enc_rpc}{Code of region, province and commune}
 \item{Fact_Pers_Reg}{Person sample weights at region level}
 \item{Fact_Pers_Com}{Person sample weights at commune level}
 \item{Fact_Hog_Reg}{Household sample weights at region level}
 \item{Fact_Hog_Com}{Household sample weights at commune level}
 \item{VarStrat}{Strata}
 \item{Conglomerado}{PSU}
 \item{VH_DV}{Households victimized by violent crimes. It works combined with Fact_Hog_*}
 \item{VH_DC}{Household victimization. It works combined  with Fact_Hog_*}
 \item{VP_DV}{People victimized by violent crimes. It works combined with Fact_Pers_*}
 \item{VP_DC}{Individual victimization. It works combined  with Fact_Pers_*}
 \item{PAD}{Perception of increased crime in the country. It works combined  with Fact_Pers_*}
 \item{rph_sexo}{1 = man; 2 = woman}
 \item{rph_edad}{Age}

 }
}
\source{
\url{https://www.ine.gob.cl/docs/default-source/seguridad-ciudadana/bbdd/2023/base-usuario-20-enusc-2023.csv?sfvrsn=34653b72_2&download=true}
}
\usage{
enusc_2023
}
\description{
ENUSC data for the year 2023. Contains only a few variables.
}
\examples{
data(enusc_2023)

}
\keyword{datasets}
