% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{epf_personas}
\alias{epf_personas}
\title{VIII Encuesta de Presupuestos Familiares}
\format{
dataframe compuesto por 48.308 observaciones y 8 variables
\describe{
 \item{sexo}{1 = male; 2 = female}
 \item{zona}{1 = metropolitan area; 2 = rest of the regional capitals}
 \item{ecivil}{marital status}
 \item{fe}{sample weights}
 \item{varunit}{PSU}
 \item{varstrat}{strata}
 \item{gastot_hd}{household expenditure}
 \item{ocupado}{1 = employed; 0 = non-employed}
}
}
\source{
\url{https://www.ine.cl/estadisticas/sociales/ingresos-y-gastos/encuesta-de-presupuestos-familiares}
}
\usage{
epf_personas
}
\description{
Reduced version of the VIII EPF database. Contains
some sociodemographic variables and the necessary information
to work with complex design.
}
\examples{
data(epf_personas)

}
\keyword{datasets}
