% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revision_calidad.R
\name{evaluar_calidad_media}
\alias{evaluar_calidad_media}
\title{Evalúa la calidad de las estimaciones de media}
\usage{
evaluar_calidad_media(tabulado, condicion = NULL, publicar = FALSE)
}
\arguments{
\item{tabulado}{\code{dataframe} generado por la función \code{crear_insumos}. Contiene
todos los insumos necesarios para la evaluación.}

\item{condicion}{character con la condición de filtro}
}
\value{
\code{dataframe} con todas las columnas que tiene el input, más una nueva que
contiene una etiqueta que da cuenta de la calidad: fiable, poco fiable o no fiable.
}
\description{
Se utiliza la metodología publicada por el INE para la evaluación de la calidad
de las estimaciones. Se consideran 3 variables: tamaño muestral, grados de libertad y
coeficiente de variación.
}
\examples{
dc <- svydesign(ids = ~varunit, strata = ~varstrat, data = epf_personas, weights = ~fe)
evaluacion_calidad(crear_insumos(gastot_hd, zona+sexo, dc))
}
