% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revision_calidad.R
\name{evaluate_mean}
\alias{evaluate_mean}
\title{Evaluate the quality of mean estimations}
\usage{
evaluate_mean(
  tabulado,
  condicion = NULL,
  publicar = FALSE,
  scheme = "chile",
  ...
)
}
\arguments{
\item{tabulado}{\code{dataframe} created by \code{crear_insumos_media}}

\item{condicion}{\code{character} with the complete condition to filter the \code{dataframe}}

\item{publicar}{\code{boolean} indicating if the evaluation of the complete table
must be added. If it is TRUE, the function adds a new column to the \code{dataframe}}
}
\value{
\code{dataframe} with all the columns included in the input table, plus a new column
containing a label indicating the evaluation of each estimation: reliable, bit reliable or unreliable
}
\description{
\code{evaluate_mean} evauates the quality of mean estimation using the
methodology created by INE Chile, which considers sample size, degrees of freedom and
coeficient of variation.
}
\examples{
dc <- survey::svydesign(ids = ~varunit, strata = ~varstrat, data = epf_personas, weights = ~fe)
evaluate_mean(create_mean(gastot_hd, dominios = zona+sexo, disenio = dc))
}
