% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revision_calidad.R
\name{evaluate_median}
\alias{evaluate_median}
\title{Evaluate the quality of median estimations}
\usage{
evaluate_median(tabulado, condicion = NULL, publicar = FALSE)
}
\arguments{
\item{tabulado}{\code{dataframe} created by \code{create_median}}

\item{condicion}{\code{character} with the complete condition to filter the \code{dataframe}}

\item{publicar}{\code{boolean} indicating if the evaluation of the complete table
must be added to the output. If it is TRUE, the function adds a new column to the \code{dataframe}}
}
\value{
\code{dataframe} with all the columns included in the input table, plus a new column
containing a label indicating the evaluation of each estimation: reliable, bit reliable or unreliable
}
\description{
\code{evaluate_median} evaluates the quality median estimations
using the methodology created by INE Chile, which considers sample size, degrees of freedom and
the coefficient of variation.
}
\examples{
dc <- svydesign(ids = ~varunit, strata = ~varstrat, data = epf_personas, weights = ~fe)
evaluate_median(crear_insumos_mediana(gastot_hd, zona+sexo, disenio =  dc))
}
