% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revision_calidad.R
\name{evaluate_prop}
\alias{evaluate_prop}
\title{Evaluate the quality of proportion estimations}
\usage{
evaluate_prop(tabulado, condicion = NULL, publicar = FALSE)
}
\arguments{
\item{tabulado}{\code{dataframe} created by \code{crear_insumos_prop}}

\item{condicion}{\code{character} with the complete condition to filter the \code{dataframe}}

\item{publicar}{\code{boolean} indicating if the evaluation of the complete table
must be added to the output. If it is TRUE, the function adds a new column to the \code{dataframe}}
}
\value{
\code{dataframe} with all the columns included in the input table, plus a new column
containing a label indicating the evaluation of each estimation: reliable, bit reliable or unreliable
}
\description{
\code{evaluar_calidad_prop} evaluates the quality of proportion estimations using the
methodology created by INE Chile, which considers sample size, degrees of freedom and
standard error.
}
\examples{
dc <- svydesign(ids = ~varunit, strata = ~varstrat, data = epf_personas, weights = ~fe)
evaluate_prop(crear_insumos_prop(ocupado, zona+sexo, dc))
}
