% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revision_calidad.R
\name{evaluate_ratio}
\alias{evaluate_ratio}
\title{Evaluate the quality of ratio estimations}
\usage{
evaluate_ratio(tabulado, condicion = NULL, publicar = FALSE)
}
\arguments{
\item{tabulado}{\code{dataframe} created by \code{create_ratio}}

\item{condicion}{\code{character} with the complete condition to filter the \code{dataframe}}

\item{publicar}{\code{boolean} indicating if the evaluation of the complete table
must be added to the output. If it is TRUE, the function adds a new column to the \code{dataframe}}
}
\value{
\code{dataframe} with all the columns included in the input table, plus a new column
containing a label indicating the evaluation of each estimation: reliable, bit reliable or unreliable
}
\description{
\code{evaluate_ratio} evaluates the quality ratio estimations
using the methodology created by INE Chile, which considers sample size, degrees of freedom, standard error and
the coefficient of variation.
}
\details{
It is evaluated according to the standard error or the coefficient of variation, based on the value of the estimation that its
receives as inputs. For estimations < 1 evaluates by standar error & for estimations > 1 evaluates by coefficient of variation.
Evaluates each estimation per row independently
}
\examples{
dc <- svydesign(ids = ~varunit, strata = ~varstrat, data = epf, weights = ~fe)
evaluate_ratio(create_ratio(var = gasto_div1, denominador = gasto, zona+sexo, dc))
}
