% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revision_calidad.R
\name{evaluate_tot_con}
\alias{evaluate_tot_con}
\title{Evaluate the quality of sum estimations for continuous variables}
\usage{
evaluate_tot_con(
  tabulado,
  condicion = NULL,
  publicar = FALSE,
  scheme = "chile",
  ...
)
}
\arguments{
\item{tabulado}{\code{dataframe} created by \code{crear_insumos_tot_con}}

\item{condicion}{\code{character} with the complete condition to filter the \code{dataframe}}

\item{publicar}{\code{boolean} indicating if the evaluation of the complete table
must be added to the output. If it is TRUE, the function adds a new column to the \code{dataframe}}
}
\value{
\code{dataframe} with all the columns included in the input table, plus a new column
containing a label indicating the evaluation of each estimation: reliable, bit reliable or unreliable
}
\description{
\code{evaluate_tot_con} evaluates the quality of total estimations for continuous variables
using the methodology created by INE Chile, which considers sample size, degrees of freedom and
the coefficient of variation.
}
\examples{


library(dplyr)
hogar <- epf_personas \%>\%
  group_by(folio) \%>\%
  slice(1)

dc <- survey::svydesign(ids = ~varunit, strata = ~varstrat, data = hogar, weights = ~fe)
evaluate_tot_con(create_tot_con(gastot_hd, dominios =  zona+sexo, disenio = dc))


}
