% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_internal.R
\name{get_cv}
\alias{get_cv}
\title{Get the coefficient of variation}
\usage{
get_cv(table, design, domains)
}
\arguments{
\item{table}{\code{dataframe} with results}

\item{design}{design}

\item{domains}{\code{vector} with domains}
}
\value{
\code{dataframe} with results including including CV
}
\description{
Receive a table created with survey and return the coefficient of variation for each cell
}
