% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_internal.R
\name{get_df}
\alias{get_df}
\title{Get degrees of freedom}
\usage{
get_df(data, domains, df_type = "eclac")
}
\arguments{
\item{data}{\code{dataframe}}

\item{domains}{\code{string} with domains}

\item{df_type}{\code{string} Use degrees of freedom calculation approach from INE Chile or eclac, by default "chile".}
}
\value{
\code{dataframe} with results including degrees of freedom
}
\description{
Receive data and domains. Returns a data frame with the psu (if available), strata and degrees of freedom (df) for each cell
}
