% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_internal.R
\name{get_sample_size}
\alias{get_sample_size}
\title{Calcula tamanio muestral para las medias}
\usage{
get_sample_size(var, data, domains, df_type = "cepal")
}
\arguments{
\item{var}{string que contiene el nombre de la variable de proporcion que se evalua.}

\item{data}{\code{dataframe} que contiene los datos que se estan evaluando}

\item{domains}{vector de caracteres que contiene los dominios a evaluar}
}
\value{
\code{dataframe} que contiene la frecuencia de todos los dominios a evaluar
}
\description{
Genera una tabla con el conteo de cada cada una de los dominios del tabulado.
La funcion contempla un caso para proporcion y un caso para promedio
}
