% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_internal.R
\name{get_survey_table}
\alias{get_survey_table}
\title{Calculates multiple estimations. Internal wrapper for survey package}
\usage{
get_survey_table(
  var,
  domains,
  complex_design,
  estimation = "mean",
  env = parent.frame(),
  fun,
  denom = NULL
)
}
\arguments{
\item{var}{\code{string} objective variable}

\item{domains}{\code{domains}}

\item{complex_design}{design from \code{survey}}

\item{estimation}{\code{string} indicating if the mean must be calculated}

\item{env}{\code{environment} parent frame}

\item{fun}{function required regarding the estimation}

\item{denom}{denominator. This parameter works for the ratio estimation}
}
\value{
\code{dataframe} containing  main results from survey
}
\description{
Generates a table with estimates for a given aggregation
}
