% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_internal.R
\name{get_table}
\alias{get_table}
\title{Calculates multiple estimations. Internal wrapper}
\usage{
get_table(
  var,
  domains,
  complex_design,
  estimation = "general",
  env = parent.frame(),
  fun,
  denom = NULL,
  type_est = "all"
)
}
\arguments{
\item{var}{\code{string} objective variable}

\item{domains}{\code{domains}}

\item{complex_design}{design from \code{survey}}

\item{estimation}{\code{string} indicating if mean, size, total, proportions, ratios must be calculated. By default 'general' specify those estimates, in the other side, median estimates are calculated with survey package.}

\item{env}{parent environment}

\item{fun}{function required regarding the estimation}

\item{denom}{denominator. This parameter works for the ratio estimation}

\item{type_est}{type of estimation: all or size}
}
\value{
\code{dataframe} containing  main results from survey
}
\description{
Generates a table with estimates for a given aggregation
}
