% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_internal.R
\name{quadratic}
\alias{quadratic}
\title{Calcula el valor de una función cuadrática}
\usage{
quadratic(p)
}
\arguments{
\item{p}{numeric vector with the values of the estimations for proportions}
}
\value{
numeric vector
}
\description{
\code{quadratic} returns the output of a particular function created by INE Chile, which
is assessed at the value of the estimated proportion from a sample. If the output of the
function is  higher than the standard error, it is interpreted as a signal that the
estimation is not reliable.
}
