% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_internal.R
\name{standardize_columns}
\alias{standardize_columns}
\title{Ordena nombre de columnas y estandariza el orden}
\usage{
standardize_columns(data, var, denom = denominador)
}
\arguments{
\item{data}{dataframe con los resultados}

\item{var}{variable objetivo}
}
\value{
dataframe con todos los datos ordenados
}
\description{
Recibe la tabla en estado bruto y la ordena
}
