% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_internal.R
\name{standardize_design_variables}
\alias{standardize_design_variables}
\title{Homologa el nombre de las variables disenio}
\usage{
standardize_design_variables(design)
}
\arguments{
\item{design}{dataframe con los resultados}
}
\value{
disenio con los nombres homologados
}
\description{
Cambia el nombre de las variables de disenio, para poder utilizarlas más adelante
}
