% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporte.R
\name{tabla_html}
\alias{tabla_html}
\title{Genera tabla en html con los resultados de la evaluaci贸n}
\usage{
tabla_html(tabla)
}
\arguments{
\item{tabla}{\code{dataframe} generado por la funciones \code{evaluacion_calidad_prop} o \code{evaluacion_calidad}.
Contiene el resultado de aplicar el protocolo de calidad.}
}
\value{
\code{html} con los resultados de la evaluaci贸n
}
\description{
La funci贸n recibe como input la evaluaci贸n de las estimaciones
}
\examples{
tabla <- evaluacion_calidad_prop(crear_insumos_prop(ocupado, zona+sexo, dc))
tabla_html(tabla)
}
