% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creacion_insumos.R
\name{unificar_variables_estrato}
\alias{unificar_variables_estrato}
\title{Homologa nombre de variable que hace referencia a los estratos de conglomerados, con el objetivo de evitar posible errores.}
\usage{
unificar_variables_estrato(disenio)
}
\arguments{
\item{disenio}{disenio complejo creado mediante el paquete \code{survey}}
}
\value{
\code{vector} que contiene la variable con los estratos de conglomerados.
}
\description{
Identifica el nombre de la variable asignada para los estratos de conglomerados en el disenio complejo, lo que permite reasignar variable con nombre estandar utilizado por las 4 funciones de creacion de insumos.
}
