% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creacion_insumos.R
\name{calcular_ic}
\alias{calcular_ic}
\title{Genera intervalos de confianza para todos los dominios estimados}
\usage{
calcular_ic(data, env = parent.frame(), tipo = "resto", ajuste_ene)
}
\arguments{
\item{data}{\code{dataframe} con todos los datos necesarios para calcular el estándar}

\item{env}{\code{environment} toma el ambiente de la función contenedora, para usar los elementos requeridos}

\item{tipo}{\code{string} que indica cuál es el tipo de estimación que se realiza.}
}
\value{
\code{dataframe} que contiene todos los elementos del estándar, junto a tres columnas nuevas que contienen el límite inferior, el límite superior y el valor t
}
\description{
Usa la tabla creada para calcular el estándar y le agrega dos columnas con el límite inferior y superior del intervalo de confianza
}
\examples{
calcular_intervalos(tabla, tipo = media_agregado)
}
