% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creacion_insumos.R
\name{calcular_tabla}
\alias{calcular_tabla}
\title{Calcula medias a partir de cierta agregación}
\usage{
calcular_tabla(var, dominios, disenio, media = T)
}
\arguments{
\item{var}{variable objetivo dentro de un \code{dataframe}. Debe anteponerse ~}

\item{dominios}{dominios de estimación separados por signo +. Debe anteponerse ~}

\item{disenio}{disenio complejo creado mediante el paquete \code{survey}}
}
\value{
\code{dataframe} que contiene variables de agregación, variable objetivo y error estándar
}
\description{
Genera una tabla con estimaciones para una agregación determinada
}
\examples{
dc <- svydesign(ids = ~varunit, strata = ~varstrat, data = epf_personas, weights = ~fe)
calcular_tabla(gastot_hd, zona+sexo, dc)

}
