% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_internal.R
\name{calcular_tabla}
\alias{calcular_tabla}
\title{Calcula medias a partir de cierta agregacion}
\usage{
calcular_tabla(var, dominios, disenio, media = T, env = parent.frame())
}
\arguments{
\item{var}{variable objetivo dentro de un \code{dataframe}. Debe anteponerse ~}

\item{dominios}{dominios de estimacion separados por signo +. Debe anteponerse ~}

\item{disenio}{disenio complejo creado mediante el paquete \code{survey}}

\item{media}{\code{boolean} indicating if the mean must be calculated}

\item{env}{\code{environment} toma el ambiente de la funcion contenedora, para usar los elementos requeridos}
}
\value{
\code{dataframe} que contiene variables de agregacion, variable objetivo y error estandar
}
\description{
Genera una tabla con estimaciones para una agregacion determinada
}
