% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creacion_insumos.R
\name{calcular_tabla_ratio}
\alias{calcular_tabla_ratio}
\title{Calcula ratio a partir de cierta agregacion}
\usage{
calcular_tabla_ratio(
  var,
  denominador,
  dominios = NULL,
  disenio,
  env = parent.frame()
)
}
\arguments{
\item{var}{variable objetivo o numerador del ratio a calcular, dentro de un \code{dataframe}. Debe anteponerse ~}

\item{denominador}{variable denominador del ratio a calcular, dentro de un \code{dataframe}. Debe anteponerse ~}

\item{dominios}{dominios de estimacion separados por signo +. Debe anteponerse ~}

\item{disenio}{disenio complejo creado mediante el paquete \code{survey}}

\item{env}{\code{environment} toma el ambiente de la funcion contenedora, para usar los elementos requeridos}
}
\value{
\code{dataframe} que contiene variables de agregacion, variable objetivo y error estandar
}
\description{
Genera una tabla con estimaciones para una agregacion determinada
}
