% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creacion_insumos.R
\name{calcular_upm}
\alias{calcular_upm}
\title{Calcula el número de UPM}
\usage{
calcular_upm(data, dominios, var = NULL)
}
\arguments{
\item{data}{\code{dataframe} que contiene los datos que se están evaluando}

\item{dominios}{vector de caracteres que contiene los dominios a evaluar}

\item{var}{string que contiene el nombre de la variable de proporción que se evalúa.}
}
\value{
\code{dataframe} que contiene la frecuencia de todos los dominios a evaluar
}
\description{
Genera una tabla con el conteo de UPM para cada uno de los dominios del tabulado.
La columna que contiene la información de las UPMs debe llamarse varunit
La función contempla un caso para proporción y un caso para promedio
}
\examples{
calcular_upm(epf_personas, c("zona", "sexo"), "ocupado")
}
