% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{casen}
\alias{casen}
\title{Encuesta de Caracterización Socioeconómica Nacional 2020 - CASEN en Pandemia 2020}
\format{
dataframe with 185.437 rows y 6 columns
\describe{
 \item{folio}{household id}
 \item{sexo}{1 = man; 2 = woman}
 \item{edad}{age}
 \item{activ}{Economic activity status}
 \item{ing_aut_hog}{Household Income}
 \item{pobreza}{poverty status: 1 = extreme poverty, 2 = non-extreme poverty, 3 = non-poverty}
 \item{expr}{regional sample weights}
 \item{estrato}{strata}
 \item{cod_upm}{PSU}
 }
}
\source{
\url{http://observatorio.ministeriodesarrollosocial.gob.cl/encuesta-casen-en-pandemia-2020}
}
\usage{
casen
}
\description{
CASEN data for the year 2020. Contains only a few variables.
}
\examples{
data(casen)

}
\keyword{datasets}
