% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_internal.R
\name{check_input_var}
\alias{check_input_var}
\title{Evalúa algunos requisitos básicos de la variable objetivo}
\usage{
check_input_var(var, disenio, estimation = "mean")
}
\arguments{
\item{var}{disenio complejo}
}
\value{
warning or stop
}
\description{
Evalúa si la variable es caracter y si es una variable de proporción en caso de que la estimación sea de media
}
