% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_internal.R
\name{chequear_var_disenio}
\alias{chequear_var_disenio}
\title{Chequea que las variables de disenio tengan el nombre correcto}
\usage{
chequear_var_disenio(data)
}
\arguments{
\item{data}{\code{dataframe} que contiene la tabla con la cual se esta trabajando}
}
\value{
un mensaje de error
}
\description{
Comprueba que las variables de disenio se llamen varstrat y varunit. En caso de que no se cumpla, la ejecucion se detiene y se genera un error
}
