% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creacion_insumos.R
\name{crear_insumos}
\alias{crear_insumos}
\title{Crea los insumos necesarios para hacer la evaluaci칩n de estimadores de la media}
\usage{
crear_insumos(var, dominios, disenio)
}
\arguments{
\item{var}{variable objetivo dentro de un \code{dataframe}. Debe anteponerse ~}

\item{dominios}{dominios de estimaci칩n separados por signo +. Debe anteponerse ~}

\item{disenio}{disenio complejo creado mediante el paquete \code{survey}}
}
\value{
\code{dataframe} que contiene la frecuencia de todos los dominios a evaluar
}
\description{
Genera una tabla con los siguientes insumos: media, grados de libertad,
tama침o muestral y coeficiente de variaci칩n.
}
\examples{
dc <- svydesign(ids = ~varunit, strata = ~varstrat, data = epf_personas, weights = ~fe)
crear_insumos(gastot_hd, zona+sexo, dc)
}
