% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creacion_insumos.R
\name{crear_insumos_mediana}
\alias{crear_insumos_mediana}
\title{Crea los insumos necesarios para hacer la evaluación de estimación de mediana}
\usage{
crear_insumos_mediana(
  var,
  dominios = NULL,
  subpop = NULL,
  disenio,
  ci = F,
  replicas = 10,
  ajuste_ene = F
)
}
\arguments{
\item{var}{variable objetivo dentro de un \code{dataframe}.}

\item{dominios}{dominios de estimación separados por signo +.}

\item{subpop}{integer dummy que permite filtrar por una subpoblación}

\item{disenio}{disenio complejo creado mediante el paquete \code{survey}}

\item{ci}{\code{boolean} que indica si los intervalos de confianza deben calcularse}

\item{replicas}{\code{integer} que indica el número de réplicas a utilizar}

\item{ajuste_ene}{\code{boolean} que indica si debe aplicarse el ajuste por cambio de marco en la ENE}
}
\value{
\code{dataframe} que contiene todos los datos del estándar
}
\description{
Genera una tabla con los siguientes insumos: mediana, grados de libertad,
tamaño muestral y error estándar. La función contempla la posibilidad de desagregar la estimación en uno o más dominios.
}
\examples{
dc <- svydesign(ids = ~varunit, strata = ~varstrat, data = epf_personas, weights = ~fe)
dc_rep <-  as.svrepdesign(dc , type = "subbootstrap", replicates=10)
crear_insumos_mediana(gastot_hd, zona+sexo, dc)
}
