% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creacion_insumos.R
\name{crear_insumos_prop}
\alias{crear_insumos_prop}
\title{Crea los insumos necesarios para hacer la evaluación de estimadores de proporción}
\usage{
crear_insumos_prop(var, dominios = NULL, subpop = NULL, disenio, ci = F)
}
\arguments{
\item{var}{variable objetivo dentro de un \code{dataframe}.}

\item{dominios}{dominios de estimación separados por signo +.}

\item{disenio}{disenio complejo creado mediante el paquete \code{survey}}
}
\value{
\code{dataframe} que contiene la frecuencia de todos los dominios a evaluar
}
\description{
Genera una tabla con los siguientes insumos: media, grados de libertad,
tamaño muestral y error estándar. La función contempla la posibilidad de desagregar la estimación en uno o más dominios.
}
\examples{
dc <- svydesign(ids = ~varunit, strata = ~varstrat, data = epf_personas, weights = ~fe)
crear_insumos_prop(ocupado, zona+sexo, dc)
}
