% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{create_html}
\alias{create_html}
\title{Genera table en html con los resultados de la evaluaci贸n}
\usage{
create_html(table)
}
\arguments{
\item{table}{\code{dataframe} generado por la funciones \code{evaluacion_label_prop} o \code{evaluacion_label}.
Contiene el resultado de aplicar el protocolo de label.}
}
\value{
\code{html} con los resultados de la evaluaci贸n
}
\description{
La funci贸n recibe como input la evaluaci贸n de las estimaciones
}
\examples{
library(survey)
library(dplyr)

hogar <- epf_personas \%>\%
  group_by(folio) \%>\%
  slice(1)
dc <- survey::svydesign(ids = ~varunit, strata = ~varstrat, data = hogar, weights = ~fe)
table <- evaluate(create_prop("ocupado", domains = "zona+sexo", design = dc))
}
