% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creacion_insumos.R
\name{create_prop}
\alias{create_prop}
\title{Create the inputs to make quality evaluation of proportion estimations}
\usage{
create_prop(
  var,
  dominios = NULL,
  subpop = NULL,
  disenio,
  ci = F,
  ajuste_ene = F
)
}
\arguments{
\item{var}{numeric variable within the  \code{dataframe}}

\item{dominios}{domains to be estimated separated by the + character.}

\item{subpop}{integer dummy variable to filter the dataframe}

\item{disenio}{complex design created by \code{survey} package}

\item{ci}{\code{boolean} indicating if the confidence intervals must be calculated}

\item{ajuste_ene}{\code{boolean} indicating if an adjustment for the sampling-frame transition period must be used}
}
\value{
\code{dataframe} that contains the inputs and all domains to be evaluated
}
\description{
\code{create_prop} generates a \code{dataframe} with the following elements: sum,
degrees of freedom, sample size and coefficient of variation and standard error. The function allows
grouping in several domains.
}
\examples{
dc <- svydesign(ids = ~varunit, strata = ~varstrat, data = epf_personas, weights = ~fe)
create_prop(ocupado, zona+sexo, disenio = dc)
}
