% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creacion_insumos.R
\name{create_prop}
\alias{create_prop}
\title{\code{create_prop} generates a \code{dataframe} with the following elements: sum,
degrees of freedom, sample size, standard error and coefficient of variation. The function allows
grouping in several domains.}
\usage{
create_prop(
  var,
  denominador = NULL,
  dominios = NULL,
  subpop = NULL,
  disenio,
  ci = F,
  ajuste_ene = F,
  standard_eval = F
)
}
\arguments{
\item{var}{numeric variable within the \code{dataframe}, is the numerator of the ratio to be calculated.}

\item{denominador}{numeric variable within the \code{dataframe}, is the denominator of the ratio to be calculated. If the \code{var} parameter is dummy, it can be NULL}

\item{dominios}{domains to be estimated separated by the + character.}

\item{subpop}{integer dummy variable to filter the dataframe}

\item{disenio}{complex design created by \code{survey} package}

\item{ci}{\code{boolean} indicating if the confidence intervals must be calculated}

\item{ajuste_ene}{\code{boolean} indicating if an adjustment for the sampling-frame transition period must be used}

\item{standard_eval}{\code{boolean} Indicating if the function is wrapped inside a function, if \code{TRUE} avoid lazy eval errors}
}
\value{
\code{dataframe} that contains the inputs and all domains to be evaluated
}
\description{
\code{create_prop} generates a \code{dataframe} with the following elements: sum,
degrees of freedom, sample size, standard error and coefficient of variation. The function allows
grouping in several domains.
}
\examples{
library(survey)
library(dplyr)
epf <- mutate(epf_personas, gasto_zona1 = if_else(zona == 1, gastot_hd, 0))
dc <- svydesign(ids = ~varunit, strata = ~varstrat, data = epf, weights = ~fe)
create_prop(var = gasto_zona1, denominador = gastot_hd, disenio =  dc)

enusc <- filter(enusc, Kish == 1)
enusc <- mutate(enusc, muj_insg_taxi = if_else(P9_4_1 \%in\% c(1,2) & rph_sexo == 2,1 ,0),
             hom_insg_taxi = if_else(P9_4_1 \%in\% c(1,2) & rph_sexo == 1,1 ,0))
dc <- svydesign(ids = ~Conglomerado, strata = ~VarStrat, data = enusc, weights = ~Fact_Pers)
options(survey.lonely.psu = "certainty")
create_prop(var = muj_insg_taxi, denominador = hom_insg_taxi, disenio = dc)


}
