% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_internal.R
\name{create_prop_internal}
\alias{create_prop_internal}
\title{internal function to calculate proportion estimations}
\usage{
create_prop_internal(
  var,
  domains = NULL,
  subpop = NULL,
  disenio,
  ci = FALSE,
  deff = FALSE,
  ess = FALSE,
  ajuste_ene = FALSE,
  rel_error = FALSE,
  log_cv = FALSE,
  unweighted = FALSE,
  standard_eval = TRUE,
  rm.na = FALSE,
  env = parent.frame()
)
}
\arguments{
\item{var}{integer dummy variable within the  \code{dataframe}}

\item{domains}{domains to be estimated separated by the + character.}

\item{subpop}{integer dummy variable to filter the dataframe}

\item{disenio}{complex design created by \code{survey} package}

\item{ci}{\code{boolean} indicating if the confidence intervals must be calculated}

\item{deff}{\code{boolean} Design effect}

\item{ess}{\code{boolean} Effective sample size}

\item{ajuste_ene}{\code{boolean} indicating if an adjustment for the sampling-frame transition period must be used}

\item{rel_error}{\code{boolean} Relative error}

\item{log_cv}{\code{boolean} indicating if the log cv must be returned}

\item{unweighted}{\code{boolean} Add non weighted count if it is required}

\item{standard_eval}{\code{boolean} indicating if the function is inside another function, by default it is TRUE, avoid problems with lazy eval.}

\item{rm.na}{\code{boolean} indicating if NA values must be removed}

\item{env}{parent environment to get some variables}
}
\value{
\code{dataframe} that contains the inputs and all domains to be evaluated
}
\description{
internal function to calculate proportion estimations
}
