% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creacion_insumos.R
\name{create_ratio_internal}
\alias{create_ratio_internal}
\title{internal function to calculate ratios estimations}
\usage{
create_ratio_internal(
  var,
  denominador,
  dominios = NULL,
  subpop = NULL,
  disenio,
  ci = F,
  ajuste_ene = F,
  anidar = T
)
}
\arguments{
\item{var}{numeric variable within the \code{dataframe}, is the numerator of the ratio to be calculated.}

\item{denominador}{numeric variable within the \code{dataframe}, is the denominator of the ratio to be calculated.}

\item{dominios}{domains to be estimated separated by the + character.}

\item{subpop}{integer dummy variable to filter the dataframe}

\item{disenio}{complex design created by \code{survey} package}

\item{ci}{\code{boolean} indicating if the confidence intervals must be calculated}

\item{ajuste_ene}{\code{boolean} indicating if an adjustment for the sampling-frame transition period must be used}

\item{anidar}{\code{boolean} indicating if the function is inside another function, by default it is TRUE, avoid problems with lazy eval.}
}
\value{
\code{dataframe} that contains the inputs and all domains to be evaluated
}
\description{
internal function to calculate ratios estimations
}
\examples{


}
