% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_internal.R
\name{create_ratio_internal}
\alias{create_ratio_internal}
\title{internal function to calculate ratios estimations}
\usage{
create_ratio_internal(
  var,
  denominator,
  domains = NULL,
  subpop = NULL,
  disenio,
  ci = FALSE,
  deff = FALSE,
  ess = FALSE,
  ajuste_ene = FALSE,
  unweighted = FALSE,
  rel_error = FALSE,
  rm.na = FALSE
)
}
\arguments{
\item{var}{numeric variable within the \code{dataframe}, is the numerator of the ratio to be calculated.}

\item{denominator}{numeric variable within the \code{dataframe}, is the denominator of the ratio to be calculated.}

\item{domains}{domains to be estimated separated by the + character.}

\item{subpop}{integer dummy variable to filter the dataframe}

\item{disenio}{complex design created by \code{survey} package}

\item{ci}{\code{boolean} indicating if the confidence intervals must be calculated}

\item{deff}{\code{boolean} Design effect}

\item{ess}{\code{boolean} Effective sample size}

\item{ajuste_ene}{\code{boolean} indicating if an adjustment for the sampling-frame transition period must be used}

\item{unweighted}{\code{boolean} Add non weighted count if it is required}

\item{rel_error}{\code{boolean} Relative error}

\item{rm.na}{\code{boolean} indicating if NA values must be removed}
}
\value{
\code{dataframe} that contains the inputs and all domains to be evaluated
}
\description{
internal function to calculate ratios estimations
}
