% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_external.R
\name{create_total}
\alias{create_total}
\title{Create the inputs to evaluate the quality of the sum of continuous variables}
\usage{
create_total(
  var,
  dominios = NULL,
  subpop = NULL,
  disenio,
  ci = F,
  ess = F,
  ajuste_ene = F,
  standard_eval = F,
  rm.na = F,
  deff = F,
  rel_error = F,
  unweighted = F
)
}
\arguments{
\item{var}{numeric variable within the  \code{dataframe}.}

\item{dominios}{domains to be estimated separated by the + character.}

\item{subpop}{integer dummy variable to filter the dataframe}

\item{disenio}{complex design created by \code{survey} package}

\item{ci}{\code{boolean} indicating if the confidence intervals must be calculated}

\item{ess}{\code{boolean} Effective sample size}

\item{ajuste_ene}{\code{boolean} indicating if an adjustment for the sampling-frame transition period must be used}

\item{standard_eval}{\code{boolean} Indicating if the function is wrapped inside a function, if \code{TRUE} avoid lazy eval errors}

\item{rm.na}{\code{boolean} Remove NA if it is required}

\item{deff}{\code{boolean} Design effect}

\item{rel_error}{\code{boolean} Relative error}

\item{unweighted}{\code{boolean} Add non weighted count if it is required}
}
\value{
\code{dataframe} that contains the inputs and all domains to be evaluated
}
\description{
\code{create_total} generates a \code{dataframe} with the following elements: sum,
degrees of freedom, sample size and coefficient of variation. The function allows
grouping in several domains.
}
\examples{
dc <- survey::svydesign(ids = ~varunit, strata = ~varstrat, data = epf_personas, weights = ~fe)
create_tot_con(gastot_hd, zona+sexo, subpop = ocupado, disenio = dc)
}
