% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_external.R
\name{create_total}
\alias{create_total}
\title{Create the inputs to evaluate the quality of the sum of continuous variables}
\usage{
create_total(
  var,
  domains = NULL,
  subpop = NULL,
  design,
  ci = FALSE,
  ess = FALSE,
  ajuste_ene = FALSE,
  standard_eval = FALSE,
  rm.na = FALSE,
  deff = FALSE,
  rel_error = FALSE,
  unweighted = FALSE,
  eclac_input = c("chile", "eclac_2020", "eclac_2023")
)
}
\arguments{
\item{var}{numeric variable within the \code{dataframe}.}

\item{domains}{domains to be estimated separated by the + character.}

\item{subpop}{integer dummy variable to filter the dataframe.}

\item{design}{complex design created by \code{survey} package.}

\item{ci}{\code{boolean} indicating if the confidence intervals must be calculated.}

\item{ess}{\code{boolean} effective sample size.}

\item{ajuste_ene}{\code{boolean} indicating if an adjustment for the sampling-frame transition period must be used.}

\item{standard_eval}{\code{boolean} indicating if the function is wrapped inside another function, if \code{TRUE} avoid lazy eval errors.}

\item{rm.na}{\code{boolean} remove NA values if required.}

\item{deff}{\code{boolean} design effect.}

\item{rel_error}{\code{boolean} relative error.}

\item{unweighted}{\code{boolean} add non-weighted count if required.}

\item{eclac_input}{\code{character} indicating the type of ECLAC inputs to return. Options are "chile", "eclac_2020", "eclac_2023". If "chile", ECLAC inputs are set to FALSE; if "eclac_2020" or "eclac_2023", ECLAC inputs are set to TRUE.}
}
\value{
\code{dataframe} that contains the inputs and all domains to be evaluated.
}
\description{
\code{create_total} generates a \code{dataframe} with the following elements: sum,
degrees of freedom, sample size, and coefficient of variation. The function allows
grouping in several domains.
}
\examples{
dc <- survey::svydesign(ids = ~varunit, strata = ~varstrat, data = epf_personas, weights = ~fe)
create_total("gastot_hd", "zona+sexo", subpop = "ocupado", design = dc)
}
