% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ene}
\alias{ene}
\title{Encuesta Nacional de Empleo - ENE. 2020-efm}
\format{
dataframe with 87.842 rows y 7 columns
\describe{
 \item{sexo}{1 = man; 2 = woman }
 \item{region}{region}
 \item{cae_especifico}{Economic activity status}
 \item{fe}{sample weights}
 \item{varunit}{PSU}
 \item{varstrat}{strata}
 \item{fdt}{It shows if the person belongs to labour force: 1 = yes; 0 = no}
 \item{ocupado}{1 = employed; 0 = non-employed}
 \item{desocupado}{1 = non-employed; 0 = employed}
}
}
\source{
\url{https://www.ine.cl/estadisticas/sociales/mercado-laboral/ocupacion-y-desocupacion}
}
\usage{
ene
}
\description{
Reduced version of the ENE database. Contains
some sociodemographic variables and the necessary information
to work with complex design
}
\examples{
data(ene)

}
\keyword{datasets}
