% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{enusc}
\alias{enusc}
\title{Encuesta Nacional Urbana de Seguridad ciudadana 2019 - ENUSC 2019}
\format{
dataframe with 24.465 rows y 22 columns
\describe{
 \item{rph_sexo}{1 = man; 2 = woman}
 \item{region}{16 regions}
 \item{Fact_Pers}{person sample weights}
 \item{Fact_Hog}{household sample weights}
 \item{Conglomerado}{PSU}
 \item{VarStrat}{strata}
 \item{VP_DC}{Individual victimization. It works combined with Fact_Pers}
 \item{VA_DC}{Household victimization. It works combined  with Fact_Hog}
 \item{rph_edad}{age}
 \item{P3_1_1}{ Perception of increased crime in the country. It works combined with Fact_Pers}
 \item{P8_1_1}{Cause of increased crime in the neighborhood. It works combined with Fact_Pers}
 \item{muj_insg_taxi}{Female perception of insecurity inside taxis. Variable elaborated with variables P9 and rph_sexo . It works combined with Fact_Pers}
 \item{hom_insg_taxi}{Male perception of insecurity inside taxis. Variable elaborated with variables P9 and rph_sexo. It works combined with Fact_Pers}
 \item{muj_insg_micro}{Female perception of insecurity inside buses. Variable elaborated with variables P9 and rph_sexo. It works combined with Fact_Pers}
 \item{hom_insg_micro}{Male perception of insecurity inside buses. Variable elaborated with variables P9 and rph_sexo. It works combined with Fact_Pers}
 \item{muj_insg_centr.com}{Female perception of insecurity inside malls. Variable elaborated with variables P9 and rph_sexo. It works combined with Fact_Pers}
 \item{hom_insg_centr.com}{Male perception of insecurity inside malls. Variable elaborated with variables P9 and rph_sexo. It works combined with Fact_Pers}
 \item{muj_insg_loc.col}{Female perception of insecurity public transport. Variable elaborated with variables P9 and rph_sexo. It works combined with Fact_Pers}
 \item{hom_insg_loc.col}{Male perception of insecurity public transport. Variable elaborated with variables P9 and rph_sexo. It works combined with Fact_Pers}
 \item{muj_insg_barrio}{Female perception of insecurity neighborhood. Variable elaborated with variables P9 and rph_sexo. It works combined with Fact_Pers}
 \item{hom_insg_barrio}{Male perception of insecurity neighborhood. Variable elaborated with variables P9 and rph_sexo. It works combined with Fact_Pers}
 }
}
\source{
\url{https://www.ine.cl/docs/default-source/seguridad-ciudadana/bbdd/2019/base-de-datos---xvi-enusc-2019-(csv).csv?sfvrsn=d3465758_2&download=true}
}
\usage{
enusc
}
\description{
ENUSC data for the year 2019. Contains only a few variables.
}
\examples{
data(enusc)

}
\keyword{datasets}
