% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revision_calidad.R
\name{evaluacion_calidad}
\alias{evaluacion_calidad}
\title{Evalúa la calidad de las estimaciones de media}
\usage{
evaluacion_calidad(tabulado)
}
\arguments{
\item{tabulado}{\code{dataframe} generado por la función \code{crear_insumos}. Contiene
todos los insumos necesarios para la evaluación.}
}
\value{
\code{dataframe} con todas las columnas que tiene el input, más una nueva que
contiene una etiqueta que da cuenta de la calidad: fiable, poco fiable o no fiable.
}
\description{
Se utiliza la metodología publicada por el INE para la evaluación de la calidad
de las estimaciones. Se consideran 3 variables: tamaño muestral, grados de libertad y
coeficiente de variación.
}
\examples{
evaluacion_calidad(crear_insumos(~gastot_hd, ~zona+sexo, dc))
}
