% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_external.R
\name{evaluate_prop}
\alias{evaluate_prop}
\title{Evaluate the quality of ratio estimations}
\usage{
evaluate_prop(
  tabulado,
  condicion = NULL,
  publicar = FALSE,
  scheme = "chile",
  ...
)
}
\arguments{
\item{tabulado}{\code{dataframe} created by \code{create_ratio}}

\item{condicion}{\code{character} with the complete condition to filter the \code{dataframe}}

\item{publicar}{\code{boolean} indicating if the evaluation of the complete table
must be added to the output. If it is TRUE, the function adds a new column to the \code{dataframe}}

\item{scheme}{string variable, default scheme is "chile" which refers to the evaluation protocol proposed by INE Chile. the alternative is "cepal" to use the CEPAL protocol}

\item{...}{the list of cepal parameters. The complete list of parameters is

1. General Parameters

\itemize{
  \item \code{df} degrees of freedom. default: 9
  \item \code{n} sample size. default ine scheme is 60. default cepal scheme: 100
}

2. INE parameters
\itemize{
  \item \code{cv_lower_ine} lower limit for cv. default: 0.15
  \item \code{cv_upper_ine} upper limit for cv. default: 0.3
}

3. CEPAL parameters
\itemize{
  \item \code{cv_cepal} limit for cv. default: 0.2
  \item \code{ess} efective sample size. default: 140
  \item \code{unweighted} unweighted count. default: 50
}}
}
\value{
\code{dataframe} with all the columns included in the input table, plus a new column
containing a label indicating the evaluation of each estimation: reliable, bit reliable or unreliable
}
\description{
\code{evaluate_prop} evaluates the quality ratio estimations
using the methodology created by INE Chile, which considers sample size, degrees of freedom, standard error and
the coefficient of variation.
}
\details{
It is evaluated according to the standard error or the coefficient of variation, based on the value of the estimation that its
receives as inputs. For estimations < 1 evaluates by standar error & for estimations > 1 evaluates by coefficient of variation.
Evaluates each estimation per row independently
}
\examples{
library(survey)
library(dplyr)
epf <-   mutate(epf_personas, gasto_zona1 = dplyr::if_else(zona == 1, gastot_hd, 0)) \%>\%
group_by(folio) \%>\%
slice(1)
dc <- svydesign(ids = ~varunit, strata = ~varstrat, data = epf, weights = ~fe)
evaluate_prop(create_prop(var = gasto_zona1, denominador = gastot_hd,
              dominios = zona+sexo, disenio = dc))
}
