% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_internal.R
\name{get_cv}
\alias{get_cv}
\title{Calcula el coeficiente de variaci칩n}
\usage{
get_cv(table, design, domains)
}
\arguments{
\item{table}{objeto creado con survey}

\item{design}{dise침o complejo creado con survey}

\item{domains}{listado de variables para desagregar}
}
\value{
dataframe con la informaci칩n de cv
}
\description{
Recibe una tabla creada con survey y devuelve el coeficiente de variaci칩n para cada celda
}
