% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_internal.R
\name{get_df}
\alias{get_df}
\title{Cálcula los grados de libertad para cada estimación}
\usage{
get_df(var, data, domains, df_type = "cepal")
}
\arguments{
\item{var}{variable objetivo}

\item{data}{dataframe}

\item{domains}{dominios en formato string}
}
\value{
dataframe con grados de libertad
}
\description{
Recibe datos y los dominios. Devuelve un data frame con las upm, varstrat y gl para cada celda
}
