% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_internal.R
\name{get_df}
\alias{get_df}
\title{Cálcula los grados de libertad para cada estimación}
\usage{
get_df(data, domains, df_type = "cepal")
}
\arguments{
\item{data}{dataframe}

\item{domains}{\code{string} with domains}

\item{df_type}{\code{string} Use degrees of freedom calculation approach from INE Chile or CEPAL, by default "ine".}
}
\value{
dataframe con grados de libertad
}
\description{
Recibe datos y los domains. Devuelve un data frame con las upm, varstrat y gl para cada celda
}
