% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_internal.R
\name{standardize_columns}
\alias{standardize_columns}
\title{standardize and sort column names}
\usage{
standardize_columns(data, var, denom)
}
\arguments{
\item{data}{\code{dataframe} with results}

\item{var}{\code{string} with the objective variable}

\item{denom}{denominator}
}
\value{
\code{dataframe} with standardized data
}
\description{
Receive the survey table in raw state and sort it
}
