% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_internal.R
\name{standardize_design_variables}
\alias{standardize_design_variables}
\title{Standardize the name of design variables}
\usage{
standardize_design_variables(design)
}
\arguments{
\item{design}{\code{dataframe}}
}
\value{
design survey
}
\description{
Rename design variables, so we can use the later
}
