% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{tabla_html}
\alias{tabla_html}
\title{Genera tabla en html con los resultados de la evaluaci贸n}
\usage{
tabla_html(tabla)
}
\arguments{
\item{tabla}{\code{dataframe} generado por la funciones \code{evaluacion_calidad_prop} o \code{evaluacion_calidad}.
Contiene el resultado de aplicar el protocolo de calidad.}
}
\value{
\code{html} con los resultados de la evaluaci贸n
}
\description{
La funci贸n recibe como input la evaluaci贸n de las estimaciones
}
\examples{
library(survey)
library(dplyr)

hogar <- epf_personas \%>\%
  group_by(folio) \%>\%
  slice(1)
dc <- survey::svydesign(ids = ~varunit, strata = ~varstrat, data = hogar, weights = ~fe)
tabla <- evaluate(create_prop("ocupado", dominios = "zona+sexo", disenio = dc))
}
